+++ tests/Documentation.test.cpp
@@ -1,7 +1,9 @@
 #include "doctest.h"
 #include "Fixture.h"
 
+#include "Luau/AstQuery.h"
 #include "LSP/DocumentationParser.hpp"
+#include "LSP/LuauExt.hpp"
 
 TEST_SUITE_BEGIN("Documentation");
 
@@ -41,6 +43,137 @@ TEST_CASE_FIXTURE(Fixture, "attach_comments_to_function_2")
     REQUIRE_EQ(0, result.errors.size());
 
     auto ty = requireType("add5");
+    auto ftv = Luau::get<Luau::FunctionType>(ty);
+    REQUIRE(ftv);
+    REQUIRE(ftv->definition);
+
+    auto comments = getCommentLocations(getMainSourceModule(), ftv->definition->definitionLocation);
+    CHECK_EQ(1, comments.size());
+}
+
+TEST_CASE_FIXTURE(Fixture, "attach_comments_to_table_property_function_1")
+{
+    auto result = check(R"(
+        local tbl = {
+            --- This is a function inside of a table
+            values = function()
+            end,
+        }
+
+        local x = tbl.values
+    )");
+
+    REQUIRE_EQ(0, result.errors.size());
+
+    auto ty = requireType("x");
+    auto ftv = Luau::get<Luau::FunctionType>(ty);
+    REQUIRE(ftv);
+    REQUIRE(ftv->definition);
+
+    auto comments = getCommentLocations(getMainSourceModule(), ftv->definition->definitionLocation);
+    CHECK_EQ(1, comments.size());
+}
+
+TEST_CASE_FIXTURE(Fixture, "attach_comments_to_table_property_function_2")
+{
+    auto result = check(R"(
+        local tbl = {
+            --- This is a function inside of a table
+            values = function()
+            end,
+            --- This is another function, and there should only be one comment connected to it
+            map = function()
+            end,
+        }
+
+        local x = tbl.map
+    )");
+
+    REQUIRE_EQ(0, result.errors.size());
+
+    auto ty = requireType("x");
+    auto ftv = Luau::get<Luau::FunctionType>(ty);
+    REQUIRE(ftv);
+    REQUIRE(ftv->definition);
+
+    auto comments = getCommentLocations(getMainSourceModule(), ftv->definition->definitionLocation);
+    CHECK_EQ(1, comments.size());
+}
+
+TEST_CASE_FIXTURE(Fixture, "attach_comments_to_props_1")
+{
+    auto result = check(R"(
+        local tbl = {
+            --- This is some special information
+            data = "hello",
+        }
+
+        local x = tbl.data
+    )");
+
+    REQUIRE_EQ(0, result.errors.size());
+
+    auto expr = Luau::findExprOrLocalAtPosition(*getMainSourceModule(), Luau::Position{6, 24}).getExpr(); // Hovering on "tbl.data"
+    REQUIRE(expr);
+
+    auto index = expr->as<Luau::AstExprIndexName>();
+    REQUIRE(index);
+
+    auto parentTyPtr = getMainModule()->astTypes.find(index->expr);
+    REQUIRE(parentTyPtr);
+    auto parentTy = Luau::follow(*parentTyPtr);
+
+    auto indexName = index->index.value;
+    auto prop = lookupProp(parentTy, indexName);
+    REQUIRE(prop);
+    REQUIRE(prop->location);
+
+    auto comments = getCommentLocations(getMainSourceModule(), prop->location.value());
+    CHECK_EQ(1, comments.size());
+}
+
+TEST_CASE_FIXTURE(Fixture, "attach_comments_to_props_2")
+{
+    auto result = check(R"(
+        --- doc comment
+        local tbl = {
+            --- This is some special information
+            data = "hello",
+        }
+
+    )");
+
+    REQUIRE_EQ(0, result.errors.size());
+
+    // Assume hovering over a var "tbl.data", which would give the position set to the property
+    auto comments = getCommentLocations(getMainSourceModule(), Luau::Location{{4, 19}, {4, 26}});
+    CHECK_EQ(1, comments.size());
+}
+
+TEST_CASE_FIXTURE(Fixture, "attach_comments_to_variable_1")
+{
+    auto result = check(R"(
+        --- doc comment
+        local var = "yo"
+    )");
+
+    REQUIRE_EQ(0, result.errors.size());
+
+    auto comments = getCommentLocations(getMainSourceModule(), Luau::Location{{2, 14}, {2, 17}});
+    CHECK_EQ(1, comments.size());
+}
+
+TEST_CASE_FIXTURE(Fixture, "attach_comments_to_variable_2")
+{
+    auto result = check(R"(
+        --- Another doc comment
+        local foo = function()
+        end
+    )");
+
+    REQUIRE_EQ(0, result.errors.size());
+
+    auto ty = requireType("foo");
     auto ftv = Luau::get<Luau::FunctionType>(ty);
     REQUIRE(ftv);
     REQUIRE(ftv->definition);
@@ -104,6 +237,65 @@ TEST_CASE_FIXTURE(Fixture, "print_multiline_comment")
             @param x number -- The number to add 5 to
             @return number -- Returns `x` with 5 added to it
         ]=]
+        function add5(x: number)
+            return x + 5
+        end
+    )");
+
+    REQUIRE_EQ(0, result.errors.size());
+
+    auto ty = requireType("add5");
+    auto ftv = Luau::get<Luau::FunctionType>(ty);
+    REQUIRE(ftv);
+    REQUIRE(ftv->definition);
+
+    auto comments = getComments(ftv->definition->definitionLocation);
+    REQUIRE_EQ(4, comments.size());
+    CHECK(comments[0] == "Adds 5 to the input number");
+    CHECK(comments[1] == "");
+    CHECK(comments[2] == "@param x number -- The number to add 5 to");
+    CHECK(comments[3] == "@return number -- Returns `x` with 5 added to it");
+}
+
+TEST_CASE_FIXTURE(Fixture, "print_multiline_comment_no_equals")
+{
+    auto result = check(R"(
+        --[[
+            Adds 5 to the input number
+
+            @param x number -- The number to add 5 to
+            @return number -- Returns `x` with 5 added to it
+        ]]
+        function add5(x: number)
+            return x + 5
+        end
+    )");
+
+    REQUIRE_EQ(0, result.errors.size());
+
+    auto ty = requireType("add5");
+    auto ftv = Luau::get<Luau::FunctionType>(ty);
+    REQUIRE(ftv);
+    REQUIRE(ftv->definition);
+
+    auto comments = getComments(ftv->definition->definitionLocation);
+    REQUIRE_EQ(4, comments.size());
+    CHECK(comments[0] == "Adds 5 to the input number");
+    CHECK(comments[1] == "");
+    CHECK(comments[2] == "@param x number -- The number to add 5 to");
+    CHECK(comments[3] == "@return number -- Returns `x` with 5 added to it");
+}
+
+
+TEST_CASE_FIXTURE(Fixture, "print_multiline_comment_variable_equals")
+{
+    auto result = check(R"(
+        --[====[
+            Adds 5 to the input number
+
+            @param x number -- The number to add 5 to
+            @return number -- Returns `x` with 5 added to it
+        ]====]
         function add5(x: number)
             return x + 5
         end